Backend/SOAP_api/soap_py/requirements.txt
requests
pandas
ipykernel
python-dotenv
lxml
beautifulsoup4>=4.12.2